Add .gitignore to exclude IDE and build files
This commit adds a .gitignore file to the root of the repository. The .gitignore file is used to specify intentionally untracked files that Git should ignore. In this case, it's now ignoring the directories ".idea/" and "cmake-build-debug/" which are auto-generated by the IDE and should not be part of the codebase.