Add 'tp3' module to project
The new tp3 module is added to extend functionality to display directory contents and metadata of specific file paths. The newly created files, "tp3.h" and "tp3.c", have been included in the CMake configuration. 'tp3.c' contains the logic to interact with directories and fetch metadata for given paths.

This feature is critical for improving file management capabilities of the application by providing more detailed file and directory info.