Remove hard-coded value for number of processes to spawn
The hard-coded value defining the number of processes to spawn was removed. This was different across the entire codebase and hence caused inconsistency. It has now been replaced with a configuration setting that can be adjusted as required. This will ensure consistency across the codebase and also enable greater flexibility in changing the number of process to spawn as per need of the program.