Improve uid and gid representation in tp3
User and Group ID in tp3 are now represented by their names instead of numbers where possible. This enhancement in tp3.c is done to improve readability and comprehension. It now handles both named and un-named IDs, defaulting back to numeric representation if no name is found for a particular UID or GID. Thus, this change enables a more informative output.