Reorganize header files and improve error message
The header files in src/main.c were slightly reorganized to improve the flow of the code. Standard library header files are now included first for better compatibility and easier debugging. Also, the error message in main function is updated to make it more clear for users if arguments are missing. In the src/tp3/tp3.c, headers were added to accommodate different platforms which increases the portability of the code.